server: Move TLS configuration out of operator
